+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.21
+
+- Add uploadImage and uploadVideo methods to Image service
+- Add searchSuggestions to Gallery service
+- Add associate method to Account service
+- Add upload_image example
+
 ## 1.0.2
 
 - Change GalleryImage accountUrl field type from int to String